chore: move styles -> old-styles (#184)
