+++ packages/chat/src/components/Bubble/styled.ts
@@ -2,7 +2,7 @@ import Button from '@/components/Button';
 import Icon from '@/components/Icon';
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 const tag = tagFactory(ClassName.BUBBLE);
