+++ packages/chat/src/components/Button/styled.ts
@@ -1,6 +1,6 @@
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 export const tag = tagFactory(ClassName.BUTTON);
