+++ packages/chat/src/components/Card/styled.ts
@@ -1,7 +1,7 @@
 import Button from '@/components/Button';
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 export const CARD_WIDTH = 246;
