+++ packages/chat/src/components/Carousel/styled.ts
@@ -2,7 +2,7 @@ import Card from '@/components/Card';
 import Icon from '@/components/Icon';
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 const BUTTON_SIZE = 42;
 export const CAROUSEL_GUTTER_WIDTH = 12;
