+++ packages/chat/src/components/Icon/styled.ts
@@ -1,6 +1,6 @@
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 const tag = tagFactory(ClassName.ICON);
