+++ packages/chat/src/components/Image/Default.tsx
@@ -1,6 +1,6 @@
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 import type { VariantProp } from '@/types';
 
 export const tag = tagFactory(ClassName.IMAGE);
