+++ packages/chat/src/components/Proactive/Close.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Icon from '@/components/Icon';
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 export const tag = tagFactory(ClassName.PROACTIVE_CLOSE);
