+++ packages/chat/src/components/Proactive/Message.tsx
@@ -1,6 +1,6 @@
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 export const tag = tagFactory(ClassName.PROACTIVE_MESSAGE);
