+++ packages/chat/src/components/Tooltip/styled.ts
@@ -2,7 +2,7 @@ import BaseButton from '@/components/Button';
 import Message from '@/components/Message';
 import { ClassName } from '@/constants';
 import { tagFactory } from '@/hocs';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 const tag = tagFactory(ClassName.TOOLTIP);
