+++ packages/chat/src/main.ts
@@ -7,7 +7,7 @@ export * from './dtos/ChatConfig.dto';
 export * from './dtos/Extension.dto';
 export * from './dtos/RenderOptions.dto';
 export * from './hooks';
-export * from './styles';
+export * from './old-styles';
 export * from './types/trace';
 export * from './types/turn';
 export * from './utils/functional';
