+++ packages/chat/src/views/ChatWindow/styled.ts
@@ -1,5 +1,5 @@
 import Chat from '@/components/Chat';
-import { styled } from '@/styles';
+import { styled } from '@/old-styles';
 
 export const ChatWindowContainer = styled('div', {
   height: '100%',
