+++ packages/chat/src/widget.tsx
@@ -5,8 +5,8 @@ import { RuntimeProvider } from './contexts';
 import type { LoadConfig } from './dtos/ChatConfig.dto';
 import { ChatConfig } from './dtos/ChatConfig.dto';
 import { RenderMode } from './dtos/RenderOptions.dto';
-import { shadowRoot } from './styles/shadow';
-import { stitches } from './styles/theme';
+import { shadowRoot } from './old-styles/shadow';
+import { stitches } from './old-styles/theme';
 import { mergeAssistantOptions } from './utils/assistant';
 import { createPlaceholderMethods } from './utils/chat';
 import { addStyleSheetURL } from './utils/stylesheet';
